refactor(4a): extract delivery loop to daemon/delivery.ts
Move defaultNotify and the daemon-direct deliveryCore/deliveryTick
into createDelivery(ctx, daemonCmux, daemonDirectCmux). cockpitd.ts
sets ctx.daemonDirectCmux and ctx.daemonCmux before calling createDelivery
so the delivery closure reads them from ctx. Re-export discoverCaptainSurface
from cockpitd.ts (test import path unchanged).
